+++ database/migrations/2023_05_11_085921_create_orders_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->string('total');
-            $table->decimal('lat');
-            $table->decimal('lng');
+            $table->double('lat');
+            $table->double('lng');
 			$table->enum('status', ['PENDING', 'PROCESSING', 'DELIVERING', 'COMPLETED', 'CANCELLED', 'REFUNDED'])->default('PENDING');
             $table->enum('payment_status', ['PENDING', 'PAID', 'FAILED'])->default('PENDING');
             $table->foreignId('buyer_id')->constrained('users', 'id')->cascadeOnDelete()->cascadeOnUpdate();
